+++ src/api/timeEntry.ts
@@ -1,5 +1,5 @@
 import { patch } from "./timecrowdClient";
 
 export const stopTimeEntry = (timeEntryId: number) => {
-  return patch(`https://timecrowd.net/api/v1/time_entries/${timeEntryId}`, {});
+  return patch(`/api/v1/time_entries/${timeEntryId}`);
 };
