+++ src/api/timecrowdClient.ts
@@ -2,12 +2,14 @@ import { getToken } from "../oauth";
 import fetch from "node-fetch";
 
 export const get = <T>(url: string) => timecrowdFetch<T>("GET", url);
-export const post = <T>(url: string, body: unknown) => timecrowdFetch<T>("POST", url, body);
-export const patch = <T>(url: string, body: unknown) => timecrowdFetch<T>("PATCH", url, body);
+export const post = <T>(url: string, body?: unknown) => timecrowdFetch<T>("POST", url, body);
+export const patch = <T>(url: string, body?: unknown) => timecrowdFetch<T>("PATCH", url, body);
+
+const baseUrl = "https://timecrowd.net";
 
 const timecrowdFetch = async <T>(method: string, url: string, body?: unknown) => {
   const token = await getToken();
-  const response = await fetch(url, {
+  const response = await fetch(`${baseUrl}${url}`, {
     method,
     headers: {
       Authorization: `Bearer ${token}`,
