+++ src/hooks/useDailyActivities.ts
@@ -1,11 +1,11 @@
 import { useCachedPromise } from "@raycast/utils";
-import { fetchDailyActivity } from "../api/dailyActivity";
+import { getDailyActivity } from "../api";
 
 import type { Dayjs } from "dayjs";
 
 export const useDailyActivities = (dates: Dayjs[]) => {
   return useCachedPromise(async () => {
-    const dailyActivities = await Promise.all(dates.map((date) => fetchDailyActivity(date)));
+    const dailyActivities = await Promise.all(dates.map((date) => getDailyActivity(date)));
 
     return dailyActivities;
   });
